Files reorganized in multiples subdir.
