+++ libui/Makefile
@@ -6,7 +6,7 @@
 #    By: alerandy <alerandy@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2019/04/05 15:11:25 by alerandy          #+#    #+#              #
-#    Updated: 2019/05/10 18:43:03 by gsmith           ###   ########.fr        #
+#    Updated: 2019/05/10 19:19:55 by gsmith           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -30,13 +30,7 @@ SDLINK = https://www.libsdl.org/release/SDL2-2.0.9.tar.gz
 SDL2 = SDL2-2.0.9
 INCLUDES += ./includes $(SDL2)/include ../libft/includes
 LIBS = -L../libft -lft -L./SDL/build/.libs -lSDL2 $(linksdl)
-SRCS += ui_init_univers.c ui_quit_univers.c ui_new_window.c ui_del_window.c \
-		ui_clear_all_windows.c ui_get_window_by_id.c ui_get_focused_window.c \
-		ui_new_event.c ui_del_event.c ui_clear_events.c ui_run_event.c \
-		ui_cmp_window.c ui_cmp_window_id.c ui_cmp_window_flag.c \
-		ui_cmp_event.c ui_cmp_event_id.c \
-		ui_free_window.c ui_free_event.c \
-		ui_generate_event_id.c  ui_pickers_a.c ui_pickers_b.c
+SRCS += ui_init_univers.c ui_quit_univers.c
 INCLUDES:=$(addprefix -I, $(INCLUDES))
 
 # Updating the VPATH
