+++ libui/src/ui_event/make.dep
@@ -1,0 +1,2 @@
+SRCS += ui_del_event.c ui_new_event.c ui_run_event.c ui_generate_event_id.c \
+		ui_clear_events.c
