+++ libui/src/ui_event/ui_del_event.c
