+++ libui/src/ui_tools/cmp/make.dep
@@ -1,0 +1,1 @@
+SRCS += ui_cmp_event.c ui_cmp_window.c
