+++ libui/src/ui_tools/cmp/ui_cmp_event.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ui_cmp_event_id.c                                  :+:      :+:    :+:   */
+/*   ui_cmp_event.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: gsmith <gsmith@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/05/06 13:52:45 by gsmith            #+#    #+#             */
-/*   Updated: 2019/05/10 11:38:32 by gsmith           ###   ########.fr       */
+/*   Created: 2019/05/06 13:45:25 by gsmith            #+#    #+#             */
+/*   Updated: 2019/05/10 19:05:31 by gsmith           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,4 +28,22 @@ int				ui_cmp_event_id(void *ptr_eve, void *ptr_id)
 	if (id_eve[0] == id_ref[0])
 		return (id_eve[1] - id_ref[1]);
 	return (id_eve[0] - id_ref[0]);
+}
+
+int			ui_cmp_event(void *ptr_a, void *ptr_b)
+{
+	unsigned int	*id_a;
+	unsigned int	*id_b;
+
+	if (!ptr_a && !ptr_b)
+		return (0);
+	if (!ptr_a)
+		return (-1);
+	if (!ptr_b)
+		return (1);
+	id_a = ((t_ui_event *)ptr_a)->id;
+	id_b = ((t_ui_event *)ptr_b)->id;
+	if (id_a[0] == id_b[0])
+		return (id_a[1] - id_b[1]);
+	return (id_a[0] - id_b[0]);
 }
