+++ libui/src/ui_tools/free/make.dep
@@ -1,0 +1,1 @@
+SRCS += ui_free_event.c ui_free_window.c
