+++ libui/src/ui_tools/pickers/make.dep
@@ -1,0 +1,1 @@
+SRCS += ui_pickers_a.c ui_pickers_b.c
