+++ libui/src/ui_window/make.dep
@@ -1,0 +1,1 @@
+SRCS += ui_new_window.c ui_del_window.c ui_get_window.c ui_clear_all_windows.c
