+++ libui/src/ui_window/ui_clear_all_windows.c
